remove non-async client
